requirements.txt
https://github.com/UB-Mannheim/tesseract/wiki
pytesseract
Pillow